generalising PrimarySliderWithAux to MainSliderWithAux with enum class RotaryType verification

+++ source/utils/math/MathConstants.h
@@ -2,7 +2,7 @@
 
 namespace particules
 {
-    static constexpr float pi = 3.141592653589793238L;
+    static constexpr float pi = static_cast<float>(3.141592653589793238L);
     static constexpr float twoPi = 2 * pi;
     static constexpr float halfPi = 0.5f * pi;
 }
